Split water_demand.rs (544 lines) into water_demand/ subdirectory
Split crates/simulation/src/water_demand.rs into focused modules
under water_demand/ for better maintainability:

- mod.rs: thin facade with re-exports
- types.rs: constants, WaterDemand component, WaterSupply resource
- calculations.rs: demand/supply calculation functions
- systems.rs: Bevy systems and WaterDemandPlugin
- tests.rs: unit tests

All public items re-exported from mod.rs — no external import changes needed.
